section 10: move remaining worker Status structs into model/status.rs
Moves ExtractionProgress (browse_view), ResizeStatus (resize_popup),
ExpandStatus (expand_hfs_dialog), and BulkConvertStatus +
BulkConvertLogLevel (bulk_convert_dialog) from per-view files into
model/status.rs. Closes section 10 items 1 + 2.

ProgressState in gui/progress.rs intentionally stays — it's UI state
(top-bar progress + controls_enabled), not worker shared-state.
